MODIStsp_check_files

Accessory function used to see if all expected out files for the selected date are already present.

MODIStsp_check_files(out_prod_folder, file_prefix, bandnames,
  bandsel_orig_choice, yy, DOY, out_format, indexes_bandnames, indexes_bandsel,
  quality_bandnames, quality_bandsel)

Arguments

out_prod_folder
string output folder name

file_prefix
string file prefix of the product (e.g., MOD13Q1)

bandnames
string array Bandnames of the MODIS product

bandsel_orig_choice
0/1 array Indicates which original modis layers were selected to be processed

yy
string year

DOY
string doy

out_format
string Gtiff or ENVI

indexes_bandnames
string array Name of available spectral indexes forthe product

indexes_bandsel
0/1 array Indicates which ospectral indexes were selected to be processed

quality_bandnames
string array Name of available Quality Indicators for the product

quality_bandsel
0/1 array Indicates which Quality Indicators were selected to be processed

Value

check - logical = 1 if all expected output files are already existing

Details

if all expected out files for the selected date are already present, check_files is set to TRUE, and date is skipped in MODIStsp_process

Note

License: GPL 3.0