MODIStsp download function

Internal function dealing with download of MODIS hdfs from http remote server for a given date.

MODIStsp_download(
  modislist,
  out_folder_mod,
  download_server,
  http,
  n_retries,
  use_aria,
  date_dir,
  year,
  DOY,
  user,
  password,
  sens_sel,
  date_name,
  gui,
  verbose
)

Arguments

modislist

character array List of MODIS images to be downloaded for the selected date (as returned from get_mod_filenames). Can be a single image, or a list of images in case different tiles are needed!

out_folder_mod

character Folder where the hdfs are to be stored

download_server

character ["http"] Server to be used.

http

character Address of the http server for the selected product.

n_retries

numeric Max number of retry attempts on download. If download fails more that n_retries times consecutively, abort

use_aria

logical if TRUE, download using aria2c

date_dir

character array Sub-folder where the different images can be found (element of the list returned from get_mod_dirs, used in case of http download to generate the download addresses).

year

character Acquisition year of the images to be downloaded

DOY

character array Acquisition doys of the images to be downloaded

user

character Username for http download

password

character Password for http download

sens_sel

character ["terra" | "aqua"] Selected sensor.

date_name

character Date of acquisition of the images to be downloaded.

gui

logical Indicates if on an interactive or non-interactive execution (only influences where the log messages are sent).

verbose

logical If FALSE, suppress processing messages, Default: TRUE

Value

The function is called for its side effects

Author

Lorenzo Busetto, phD (2014-2017)

Luigi Ranghetti, phD (2015) luigi@ranghetti.info