Retrieve the names of MODIS layers for a product

Function used to retrieve the names of original MODIS layers, quality layers and eventually available spectral indexes given a MODIS product code. It is useful to identify the names of the layers to be processed when launching MODIStsp from the CLI.

MODIStsp_get_prodlayers(prodname)

Arguments

prodname
character containing the code of the desired MODIS product. NOTE: for products available separately for Terra and Aqua (e.g., MOD13Q1/MYD13Q1), use the code MD_code_ (e.g., MD13Q1)

Value

list, containing the slots: prodname, bandnames, quality_bandnames and indexes_bandnames, band_fullnames, quality_fullnames, indexes_fullnames

Note

License: GPL 3.0

Author

Lorenzo Busetto, phD (2014-2020)

Examples

# Get layers of product M*13Q1 based on code
MODIStsp_get_prodlayers("M*13Q1")

#> $prodname
#> [1] "Vegetation Indexes_16Days_250m (M*D13Q1)"
#> 
#> $bandnames
#>  [1] "NDVI"     "EVI"      "VI_QA"    "b1_Red"   "b2_NIR"   "b3_Blue" 
#>  [7] "b7_SWIR"  "View_Zen" "Sun_Zen"  "Rel_Az"   "DOY"      "Rely"    
#> 
#> $bandfullnames
#>  [1] "16 day NDVI average"                "16 day EVI average"                
#>  [3] "VI quality indicators"              "Surface Reflectance Band 1"        
#>  [5] "Surface Reflectance Band 2"         "Surface Reflectance Band 3"        
#>  [7] "Surface Reflectance Band 7"         "View zenith angle of VI pixel"     
#>  [9] "Sun zenith angle of VI pixel"       "Relative azimuth angle of VI pixel"
#> [11] "Day of year of VI pixel"            "Quality reliability of VI pixel"   
#> 
#> $quality_bandnames
#> [1] "QA_qual"     "QA_usef"     "QA_aer"      "QA_adj_cld"  "QA_BRDF"    
#> [6] "QA_mix_cld"  "QA_land_wat" "QA_snow_ice" "QA_shd"     
#> 
#> $quality_fullnames
#> [1] "VI Quality"                          
#> [2] "VI usefulness"                       
#> [3] "Aerosol quantity"                    
#> [4] "Adjacent cloud detected"             
#> [5] "Atmosphere BRDF correction performed"
#> [6] "Mixed Clouds"                        
#> [7] "Land/Water Flag"                     
#> [8] "Possible snow/ice"                   
#> [9] "Possible shadow"                     
#> 
#> $indexes_bandnames
#> [1] "SR"    "NDFI"  "NDII7" "SAVI" 
#> 
#> $indexes_fullnames
#> [1] "Simple Ratio (NIR/Red)"              
#> [2] "Flood Index (Red-SWIR2)/(Red+SWIR2)" 
#> [3] "NDII7 (NIR-SWIR2)/(NIR+SWIR2)"       
#> [4] "SAVI (NIR-Red)/(NIR+Red+0.5)*(1+0.5)"
#> 

# Get layers of product M*13Q1 based on full name
MODIStsp_get_prodlayers("Vegetation Indexes_16Days_250m (M*D13Q1)")

#> $prodname
#> [1] "Vegetation Indexes_16Days_250m (M*D13Q1)"
#> 
#> $bandnames
#>  [1] "NDVI"     "EVI"      "VI_QA"    "b1_Red"   "b2_NIR"   "b3_Blue" 
#>  [7] "b7_SWIR"  "View_Zen" "Sun_Zen"  "Rel_Az"   "DOY"      "Rely"    
#> 
#> $bandfullnames
#>  [1] "16 day NDVI average"                "16 day EVI average"                
#>  [3] "VI quality indicators"              "Surface Reflectance Band 1"        
#>  [5] "Surface Reflectance Band 2"         "Surface Reflectance Band 3"        
#>  [7] "Surface Reflectance Band 7"         "View zenith angle of VI pixel"     
#>  [9] "Sun zenith angle of VI pixel"       "Relative azimuth angle of VI pixel"
#> [11] "Day of year of VI pixel"            "Quality reliability of VI pixel"   
#> 
#> $quality_bandnames
#> [1] "QA_qual"     "QA_usef"     "QA_aer"      "QA_adj_cld"  "QA_BRDF"    
#> [6] "QA_mix_cld"  "QA_land_wat" "QA_snow_ice" "QA_shd"     
#> 
#> $quality_fullnames
#> [1] "VI Quality"                          
#> [2] "VI usefulness"                       
#> [3] "Aerosol quantity"                    
#> [4] "Adjacent cloud detected"             
#> [5] "Atmosphere BRDF correction performed"
#> [6] "Mixed Clouds"                        
#> [7] "Land/Water Flag"                     
#> [8] "Possible snow/ice"                   
#> [9] "Possible shadow"                     
#> 
#> $indexes_bandnames
#> [1] "SR"    "NDFI"  "NDII7" "SAVI" 
#> 
#> $indexes_fullnames
#> [1] "Simple Ratio (NIR/Red)"              
#> [2] "Flood Index (Red-SWIR2)/(Red+SWIR2)" 
#> [3] "NDII7 (NIR-SWIR2)/(NIR+SWIR2)"       
#> [4] "SAVI (NIR-Red)/(NIR+Red+0.5)*(1+0.5)"
#> 

# Get indexes names of product M*13Q1 based on full name
MODIStsp_get_prodlayers("MCD43C4")$indexes_bandnames

#>  [1] "NDVI"  "EVI"   "SR"    "NDFI"  "NDII6" "NDII7" "SAVI"  "NDSI"  "GNDVI"
#> [10] "RGRI"  "GRVI"