Read MODIS products characteristics from XML

function used to parse the XML file used to store the characteristics of MODIS Land Products and store them in the "prod_opts" data frame

MODIStsp_read_xml(prodopts_file, xml_file)

Arguments

prodopts_file
string filename of the RData in which to store the data parsed from the XML file

xml_file
string filename of the XML file containing the MODIS products characteristics

Value

NULL - retrieved data are stored in the specified RData file

Details

The function parses the XML file product by product, stores data in a data frame and saves the data frame within the "MODIStsp_previous" RData file as a list of lists

Note

License: GPL 3.0