Create MODIStsp virtual files

Function used to create virtual files from time series of single-band files corresponding to different acquisition dates. The function takes as input the folder in which the single-band files are stored, and creates a ENVI Meta file and/or a GDAL vrt file that allows access to the full time series as if it was a single physical file. Created virtual files are stored in the "Time Series" subfolder of `out_prod_folder``

Usage

MODIStsp_vrt_create(
  sensor,
  out_prod_folder,
  bandnames,
  bandsel,
  nodata_out,
  indexes_bandnames,
  indexes_bandsel,
  indexes_nodata_out,
  quality_bandnames,
  quality_bandsel,
  quality_nodata_out,
  file_prefixes,
  ts_format,
  out_format,
  verbose
)

Arguments

sensor

character ["Terra"| "Aqua" | "Both"] MODIS platform to be considered. (Ignored for MCD* products). Default: "Both"

out_prod_folder

character Main output folder.

bandnames

names of all layers available for the product being processed

bandsel

character array Original MODIS layers to be processed. You can get a list of available layers for a given product using function MODIStsp_get_prodlayers (e.g., MODIStsp_get_prodlayers("M*D13Q1")$bandnames), Default: NULL

nodata_out

numeric Output nodata value to be used in vrt files

indexes_bandnames

names of all indexes available for the product being processed

indexes_bandsel

character arraySpectral Indexes to be computed starting from reflectance bands. You can get a list of available quality layers for a given product using function MODIStsp_get_prodlayers (e.g., MODIStsp_get_prodlayers("M*D13Q1")$indexes_bandnames), Default: NULL

indexes_nodata_out

nodata value for indexes vrts

quality_bandnames

names of all quality indicators available for the product being processed

quality_bandsel

character array Quality Indicators to be computed starting from bit fields of original MODIS layers. You can get a list of available quality layers for a given product using function MODIStsp_get_prodlayers (e.g., MODIStsp_get_prodlayers("M*D13Q1")$quality_bandnames), Default: NULL

quality_nodata_out

nodata value for quality vrts

file_prefixes

character array (2) file_prefixes for TERRA and AQUA - used to identify the files corresponding to each sensor

ts_format

character ["ENVI" | "GDAL" | "Both"] Required output format for virtual file.

out_format

character ["ENVI" | "GTiff"] Format of images used as "input" for the vrt and contained in out_prod_folder/band folders.

verbose

logical If FALSE, suppress processing messages, Default: TRUE

Value

NULL - the function is called for its side effects

Note

License: GPL 3.0

Author

Lorenzo Busetto, phD (2014-2017)

Luigi Ranghetti, phD (2015)