Create template code for modules

xxx

This module contains the configuration informations for the corresponding pipeline. It is called by the nav_pipeline module of the package MagellanNTK

The name of the server and ui functions are formatted with keywords separated by '_', as follows: * first string `mod`: indicates that it is a Shiny module * `pipeline name` is the name of the pipeline to which the process belongs * `process name` is the name of the process itself

This convention is important because MagellanNTK call the different server and ui functions by building dynamically their name.

In this example, `#name#_ui()` and `#name#_server()` define the code for the process `xxx` which is part of the pipeline called `xxx`.

This function inserts the necessary code for the 'Description' step

This function inserts the necessary code for the 'Description' step

This function inserts the necessary code for the Description' step

createModuleTemplate(config = NULL, path = ".")

write_general_comment(con, name)

write_ui_func(con, name)

write_header_server_func(con, name)

write_config_func(con, config)

write_process_code_for_default_value_widgets(con)

write_module_server_header(con, mode)

write_process_renderUI_for_steps(con, ll.config, path)

write_stepUI_template(step.name)

write_Insert_Description_Step_code_for_Process(con, path)

write_Insert_Save_Step_code_for_Process(con)

Create_Pipeline_Description_source_file(fullname, path)

write_output_func(con)

vec2code(ls_list, is.char = FALSE)

Arguments

config

xxx

path

xxx

con

xxx

name

xxx

mode

xxx

ll.config

xxx

step.name

xxx

fullname

xxx

ls_list

A vector

is.char

A `bolean(1)` to indicate whether the items are strings or not. In this case, they will be quoted in the result

id

xxx

dataIn

The dataset

steps.enabled

A vector of boolean which has the same length of the steps of the pipeline. This information is used to enable/disable the widgets. It is not a communication variable between the caller and this module, thus there is no corresponding output variable

remoteReset

It is a remote command to reset the module. A boolean that indicates is the pipeline has been reseted by a program of higher level Basically, it is the program which has called this module

steps.status

xxx

current.pos

xxx

Value

A string