Estimate the expected variance of beta

Estimate the expected variance of beta. This is approximately expected(1/var(U)).

expected_vbeta(N0, N1, snps, W, gamma.W, freq, GenoProbList)

Arguments

N0
The number of Y=0

N1
The number of Y=1

snps
The snps at which we wish to compute the expected Z Score

W
The true causal SNPs (these need not be in "snps")

gamma.W
The log odds ratios of effect of the true causal SNPs (not including gamma0, the intercept term)

freq
Haplotype frequencies as a data.frame, with column Probability indicating relative frequency in controls.

GenoProbList
An list of objects giving the probability of seeing each X,W genotype vector. This can be calculated within the function if no value supplied, or you can pass a pre-calculated version

Value

The expected variance of beta for each SNP X, assuming the causal SNPs are W

Details

Assumes we have a list, GenoProbList, giving the GenoProb values for each X.