Find MODIS dates included in selected processing period

Accessory function to find the folders corresponding to the requested dates period within the full list retrieved by get_moddirs

get_mod_dates(dates, date_dirs)

Arguments

dates
2- element string array specifying start/end dates (yyyy.mm.dd) for which the http addresses of folders in lpdaac should be retrieved (e.g., c("2015.1.1", "2015.12.31)

date_dirs
data frame full list of folders in lpdaac archive for product of interest

Value

array of folder names containing data for the MODIS product acquired in the period specified by "dates"

Note

License: GPL 3.0