identify dates to be processed for a year
Source: R/get_yeardates.R

helper function needed to identify the ranges of dates to be processed for a given year as a function of download_range selection and starting/ending dates and years

get_yeardates(download_range, yy, start_year, end_year, start_date, end_date)

Arguments

download_range
character ["full" | "seasonal"] If "full", all the available images between the starting and the ending dates are downloaded; If "seasonal", only the images included in the season are downloaded (e.g: if the starting date is 2005-12-01 and the ending is 2010-02-31, only the images of December, January and February from 2005 to 2010 - excluding 2005-01, 2005-02 and 2010-12 - are downloaded).

yy
numeric year for which the processing dates need to be identified

start_year
numeric start year of current MODIStsp_process run

end_year
numeric end year of current MODIStsp_process run.

start_date
character`` Start date for images download and preprocessing (yyyy.mm.dd) of currentMODIStsp_process` run.

end_date
character Start date for images download and preprocessing (yyyy.mm.dd) of current MODIStsp_process run.

Value

OUTPUT_DESCRIPTION