Load MODIStsp processing options from a JSON file

Load MODIStsp processing option from opts_jsfile if it already exist, otherwise initialize processing options to default and save them to opts_jsfile (typically done at first execution, or if the MODIStsp_previous.json is deleted). Sends warnings if options file is from an old version. Aborts if the json file is not a valid MODIStsp options file

load_opts(opts_jsfile)

Arguments

opts_jsfile
Expected file name of the JSON file containing processing options

Value

data frame general_opts, containing the processing options retrieved from the JSON file (or the defaults set at first execution). See also MODIStsp_GUI and MODIStsp_process

Note

License: GPL 3.0