lpdaac_getmod_dirs

Accessory function to get the full list of directories on the lpdaac http site (modified after Barry Rowlingson function):

lpdaac_getmod_dirs(ftp, http, used_server = NA, user = user,
  password = password, gui, out_folder_mod, .Platform)

Arguments

ftp
string ftp site corresponding to a given MODIS product

http
string http site on lpdaac corresponding to a given MODIS product

used_server
string can assume values "http" or "ftp" depending on the used download server; if NA, the script tries to download with http, using ftp if the download fails

user
username for earthdata server

password
password for earthdata server

gui
logical indicates if processing was called within the GUI environment or not. If not, direct processing messages to the log

out_folder_mod
ouput folder for original HDF storage

.Platform
string os platform (from call to .Platform)

Value

list of all available folders (a.k.a. dates) for the requested MODIS product on lpdaac archive

Note

License: GPL 3.0