Load customizable functions

A shiny Module.which allow to choose whether to use the default functions (embedded in Magellan) or functions provided by external packages. The names of these functions (shiny modules) are: * convert(): xxx * open_dataset(): xxx * open_demoDataset(): xxx * view_dataset(): xxxx * infos_dataset(): xxx

Thus, MagellanNTK seraches in the global environment if any package exports one or more of these functions. If so,

For each of these functions, the shiny app lists all the packages that export it. Once the user has make its choices, the module returns a list containing the infos.

This code is extracted from https://sebastiansauer.github.io/finds_funs/

loadapp_ui(id)

loadapp_server(id)

find_funs(f)

loadApp()

Arguments

id

shiny id

f

name of function for which the package(s) are to be identified.

input

internal

output

internal

session

internal

Value

A list where each slot is named with the customizable functions. Each slot contains the code to call this function from the package the user have chosen.

A dataframe with two columns:

Examples

if (FALSE) { # interactive()
app <- loadApp()
shiny::runApp(app)

}
find_funs('filter')
#> # A tibble: 10 × 3
#>    package_name    builtin_pckage loaded
#>    <chr>           <lgl>          <lgl> 
#>  1 BiocGenerics    FALSE          FALSE 
#>  2 base            TRUE           TRUE  
#>  3 clusterProfiler FALSE          FALSE 
#>  4 dplyr           FALSE          TRUE  
#>  5 mice            FALSE          FALSE 
#>  6 plotly          FALSE          FALSE 
#>  7 rstatix         FALSE          FALSE 
#>  8 stats           TRUE           TRUE  
#>  9 tidygraph       FALSE          FALSE 
#> 10 tidytree        FALSE          FALSE