Spawn processing update messages

helper function to provide processing messages

process_message(mess_text, gui, mess_lab, verbose = TRUE)

Arguments

mess_text
character text to be shown in the processing windows and/or the console

gui
logical indicating if the message should be passed to the status window or only to the console.

mess_lab
pointer to the gwindow used to shoe messages when gui == TRUE

verbose
logical If FALSE, suppress processing messages, Default: TRUE

Value

The function is called for its side effects