MODIStsp

Main function for the MODIS Time Series Processing Tool (MODIStsp)

MODIStsp(gui = TRUE, options_file = NULL, spatial_file_path = NULL,
  scroll_window = FALSE, test = -1, n_retries = 20)

Arguments

gui
logical if TRUE: the GUI is opened before processing. If FALSE: processing parameters are retrieved from the provided options_file argument), Default: TRUE

options_file
character full path to a JSON file containing MODIStsp processing options saved from the GUI. If NULL, parameters of the last successful run are retrieved from file "MODIStsp_Previous.json" in subfolder "Previous"), Default: NULL

spatial_file_path
character (optional) full path of a spatial file to use to derive the processing extent. If not NULL, the processing options which define the extent, the selected tiles and the "Full Tile / Resized" in the JSON options file are overwritten and new files are created on the extent of the provided spatial file, Default: NULL

scroll_window
logical if TRUE, the GUI window is opened fullscreen with scrollbars (this is useful on devices with small displays). If using a device with a display resolution >= 1024x768, leaving this parameter to FALSE is suggested, Default: FALSE

test
integer if set, MODIStsp is executed in "test mode", using a preset Options File instead than opening the GUI or accepting the options_file parameter. This allows both to check correct installation on user's machines, and to implement unit testing. The number indicates which tests to execute (six are available). If == 0, the test is selected randomly. If == -1, MODIStsp is executed in normal mode. Default: -1

n_retries
numeric maximum number of retries on download functions. In case any download function fails more than n_retries times consecutively, MODIStsp_process will abort, Default: 20

Details

The function is used to:

initialize the processing (folder names, packages, etc.);

launch the GUI (MODIStsp_GUI()) and receive its outputs on interactive execution, or load an options file on non-interactive execution;

launch the routines for downloading and processing the requested datasets. (MODIStsp_process())

Note

License: GPL 3.0

See also

MODIStsp_GUI(), MODIStsp_process()

Examples

# Running the tool without any option will start the GUI with the default or
# last used settings

# NOT RUN {
MODIStsp()
# }

# Running the tool using the settings previously saved in a specific options
# file

# NOT RUN {
MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json")
# }

# Running the tool using the settings previously saved in a specific option file
# and specifying the extent from a spatial file allows to re-use the same 
# processing settings to perform download and reprocessing on a different area

# NOT RUN {
MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
  spatial_file_path = "X:/yourpath/yourspatialfile.shp" )
# }

# Runnig the tool using the settings previously saved in a
# specific options file and specifying each time the extent from a different
# spatial file (e.g., to perform the same processing on several extents)

# NOT RUN {
extent_list = list.files("X:/path/containing/some/shapefiles/", "\\.shp$")
for (single_shape in extent_list)
  MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
    spatial_file_path = single_shape )
# }
    
(See http://lbusett.github.io/MODIStsp/articles for additional details and 
examples)

#> Error: <text>:2:6: unexpected symbol
#> 1:     
#> 2: (See http
#>         ^