Build and manage the MODIStsp GUI

Function used to generate and handle the GUI used to allow selection of MODIStsp processing parameters. If the "previous options" file (MODIStsp_Previous.json) already exists, it is loaded and used to reinstate the GUI to its last state. Otherwise, the previous options file is created by launching the MODIStsp_read_xml function

MODIStsp_GUI(general_opts, prod_opt_list, MODIStsp_dir, opts_jsfile,
  prodopts_file, scroll_window)

Arguments

general_opts
data.frame containing general processing options passed by MODIStsp

prod_opt_list
List of MODIS products specifications (read from MODIStsp_ProdOpts.xml file)

MODIStsp_dir
main folder of the package

opts_jsfile
json parameters file containing data of the last execution, or the ones contained in the options_file eventually passed to MODIStsp

prodopts_file
rdata file containing info about MODIS products

scroll_window
logical parameter passed by MODIStsp main function.

Value

start - Logical - tells the main if running processing or exiting (also, Processing options are saved in "previous" file and (if "Save options" is pressed) in user's selected file)

Note

License: GPL 3.0