Add custom spectral indexes

Function used to add a user-defined Spectral Index to the default list of computable spectral indexes. Execution without the GUI (i.e., to add a new index from a script) is also possible (see examples).

MODIStsp_addindex(opts_jsfile = NULL, prodopts_file = NULL,
  selprod = NULL, selvers = NULL, gui = TRUE, new_indexbandname = "",
  new_indexfullname = "", new_indexformula = "",
  new_indexnodata_out = "32767", MODIStsp_dir = system.file(package =
  "MODIStsp"))

Arguments

opts_jsfile
character full path of a JSON file containing the processing options in which the new indexes has to be saved (default: MODIStsp_Previous.JSON in subfolder Previous).

prodopts_file
character: full path of the RData file containing. if NULL, use MODIStsp_ProdOpts.RData in subfolder Previous, Default: NULL

selprod
character Name of the product to which the new index should be added (Note: the index will be added to all other products allowing its computation !). If NULL, as in non-interactive execution, no check on available band names is skipped and the index (if valid) is added to all products supporting it, Default: NULL

selvers
character Version of the product to which the new index should be added (Note: the index will be added to all other products allowing its computation !). If NULL, as in non-interactive execution, the check on available band names is skipped and the index (if valid) is added to all products supporting it, Default: NULL

gui
logical if TRUE, a GUI is opened to define the new index; otherwise use the "new_indexbandname", "new_indexfullname" and "new_indexformula" parameters to define it non-interactively, Default: TRUE

new_indexbandname
character short name (acronym) of the new spectral index (Ignored if gui == TRUE), Default: NULL

new_indexfullname
character extended name (acronym) of the new spectral index (Ignored if gui == TRUE), Default: NULL

new_indexformula
character string containing the formula of the new spectral indexes (Ignored if gui == TRUE). Variables allowed in the formula are the names of the bands: b1_Red, b2_NIR, b3_Blue, b4_Green, b5_SWIR, b6_SWIR and b7_SWIR. Default: NULL

new_indexnodata_out
character nodata value to use for rasters containing the new index

MODIStsp_dir
character main folder containing MODIStsp R files, Default: retrieved from package installation folder

Value

The function is called for its side effects. On success, the MODIStsp_Previous.RData file is modified so to allow computation of the additional indexes.

Details

The function asks the user to provide the info related to the new desired Spectral Index using a GUI interface, checks for correctness of provided information (e.g., correct bandnames, computable formula, etc...). If the index is legit, it modifies the MODIStsp_Previous.json (or of the json file provided by the user) so to allow computation of the additional index within MODIStsp.

To remove all custom-added spectral indexes, simply delete the MODIStsp_Previous.json file within the /Previous subfolder of the folder in which the package was installed, or the alternative JSON specified by the parameter "opts_jsfile".

The function can be run either from within the main MODIStsp GUI, or within a stand-alone script (using GUI = FALSE). In the latter case, it modifies either the MODIStsp_Previous.RData options file, or the options_file specified by the user to add the new index, without user interaction.

Note

License: GPL 3.0

See also

MODIStsp_resetindexes

Examples

# Run the GUI to interactively define a new index
 
# NOT RUN {
 MODIStsp_addindex()
# }

# Open the GUI to define a new index and save it in a custom json
# options file.

# NOT RUN {
opts_jsfile = system.file("testdata/test_addindex.json", package = "MODIStsp")
MODIStsp_addindex(opts_jsfile = opts_jsfile)
# }

# Define the new index in non-interactive execution, without specifying an
# options file (thus modifying MODIStsp_previous.json)

# NOT RUN {
MODIStsp_addindex(gui = FALSE, new_indexbandname = "SSI",
  new_indexfullname = "Simple Useless Index",
  new_indexformula = "b2_NIR+b1_Red")
# }