MODIStsp_process_QA_bits

function used to extract quality indicator from MODIS aggregated quality layers

MODIStsp_process_QA_bits(out_filename, in_raster_name, bitN, source,
  out_prod_folder, file_prefix, yy, DOY, out_format, nodata_source,
  nodata_qa_in, nodata_qa_out)

Arguments

out_filename
string file name of the output raster files containing QI values

in_raster_name
name of the MODIS band containing data from which the bit field corresponding to the quality indicator must be extracted

bitN
position of the bits corresponding to the quality indicator (e.g., 0-1 = first two bits; 2-5: bits from 2 to 5, etc.)

source
"name" of the MODIS band containing the values from which to extract the quality indicator (e.g., State_1km)

out_prod_folder
main folder used for storing the output data in MODIStsp processing. New quality indicator files are stored in out_prod_folder/derived_band

file_prefix
files_prefixes used to create output file names in MODIStsp (e.g., MOD13Q1)

yy
year of considered image

DOY
DOY of considered image

out_format
output format (ENVI or GTiff)

nodata_source
nodata values of the MODIS band containing data from which the bit field corresponding to the quality indicator must be extracted

nodata_qa_in
string in nodata for quality bands ("255")

nodata_qa_out
string out nodata for quality bands ("255")

Details

On the basis of the name of the image containing the aggregated quality information (in_raster_name) and of the position of the bit fields corresponding to the QI of interest in the bitfield representation (bitN), the function extracts the correct information exploiting bitwise operators, and save the result in a new raster image

Note

License: GPL 3.0 Based on the "modis.qc.R" script by Yann Chemin (2008) (https://r-forge.r-project.org/scm/viewvc.php/pkg/RemoteSensing/R/modis.qc.R?view=markup&root=remotesensing&pathrev=79)

license GPL 3.0