MODIStsp helper for processing original HDF layers

Internal function used to performed the required spatial processing on MODIS original hdf layers (reprojection, resizing, resampling, mosaicing, computation of scaling factors). The function is based on the use of gdal routines.

MODIStsp_process_bands(out_folder_mod, modislist, outproj_str, mod_proj_str,
  sens_sel, band, bandname, date_name, datatype, nodata_in, nodata_out,
  full_ext, bbox, scale_val, scale_factor, offset, out_format, outrep_file,
  compress, out_res_sel, out_res, resampling, gui, mess_lab, verbose)

Arguments

out_folder_mod
character Output folder for original HDF storage.

modislist
character array List of MODIS images to be downloaded for the selected date (as returned from get_mod_filenames). Can be a single image, or a list of images in case different tiles are needed!

outproj_str
character proj4 string of selected output projection.

mod_proj_str
character proj4 string of MODIS product native projection.

sens_sel
character ["terra" | "aqua"] Selected sensor.

band
numeric band number corresponding to the HDF layer to be processed

bandname
character Name to the HDF layer to be processed.

date_name
character Date of acquisition of the images to be downloaded.

datatype
character Datatype to the HDF layer to be processed.

nodata_in
numeric Original nodata value to the HDF layer to be processed.

nodata_out
numeric Output nodata value to the HDF layer to be processed.

full_ext
characrter ["Full_Ext" | "Resized"] If "Full_Ext", process the entire extent of the selected tiles. Otherwise, crop the output to output bbox.

bbox
numeric(4) Output bounding box (xmin, xmax, ymin, ymax) in out_proj coordinate system.

scale_val
character ["Yes" | "No"] If == "Yes", scale and offset are applied to original MODIS layers, and Spectral Indexes are saved as floating point. If == "No", no rescaling is done and Spectral Indexes are saved as integer, with a 10000 scaling factor.

scale_factor
numeric Scale factor to be applied to the HDF layer to be processed (Ignored if scale_val == FALSE).

offset
numeric Offset to be applied to the HDF layer to be processed (Ignored if scale_val == FALSE).

out_format
character ["ENVI" | "GTiff"] Desired output format.

outrep_file
character Full path of the file where results of the processing are to be stored (created in MODIStsp_process)

compress
character ["None" | "PACKBITS" | "LZW" | "DEFLATE"] Compression method for GTiff outputs (Ignored if out_format == ENVI)

out_res_sel
character ["Native" | "Resampled"] Indicates if the native resolution of the product or a user supplied one is to be used.

out_res
float Output resolution (in output projection measurement unit). Ignored if out_res_sel == "Native".

resampling
character ["near" | "bilinear" | "cubic" | "lanczos" | "mode"] Resampling method to be used by gdalwarp.

gui
logical Indicates if processing was called starting from an interactive environment or not. If FALSE, processing messages are sent to a log file instead than to the console, and gWidgets messages are suppressed

mess_lab
Pointer to the gWidget used to visualize processing messages in interactive execution.

verbose
logical If FALSE, suppress processing messages, Default: TRUE

Value

The function is called for its side effects