MODIStsp_reset_options

Helper function used to reset MODIStsp options to default by removing the MODIStsp_Previous.json file. May be useful to get back to a "working" state if the GUI gets somehow corrupted due to an invalid MODIStsp_Previous.json file used.

MODIStsp_reset_options()

Value

The function is called for its side effects