Create MODIStsp virtual files

Function used to create virtual files from time series of single-band files corresponding to different acquisition dates. The function takes as input the folder in which the single-band files are stored, and creates a ENVI Meta file and/or a GDAL vrt file that allows access to the full time series as if it was a single physical file. Created virtual files are stored in the "Time Series" subfolder of `out_prod_folder``

MODIStsp_vrt_create(sensor, out_prod_folder, bandnames, bandsel, nodata_out,
  indexes_bandnames, indexes_bandsel, indexes_nodata_out, quality_bandnames,
  quality_bandsel, quality_nodata_out, file_prefixes, ts_format, out_format,
  verbose)

Arguments

sensor
character ["Terra"| "Aqua" | "Both"] MODIS platform to be considered. (Ignored for MCD* products).

out_prod_folder
character Main output folder.

bandnames
character array Abbreviated Names of original layers of the selected product (used to build output file names).

bandsel
integer 0/1 array of length equal to number of original layers of the selected product set to 1 for bands to be processed.

nodata_out
numeric array Target NoData values of MODIS original layers (Ignored if nodata_change == FALSE).

indexes_bandnames
character array Abbreviated Names of SIs available for the selected product (used to build output file names of SIs).

indexes_bandsel
integer 0/1 array array of length equal to the number of Spectral Indexes available for the product (standard + user-provided), set to 1 for indexes to be processed.

indexes_nodata_out
numeric array NoData values to be used for SIs

quality_bandnames
character array Abbreviated Names of Quality Indicators available for the selected product (used to build output file names of QIs).

quality_bandsel
integer 0|1 array array of length equal to number of available QIs, set to 1 for indexes to be processed.

quality_nodata_out
integer Always set to 255.

file_prefixes
character array (2) file_prefixes for TERRA and AQUA - used to identify the files corresponding to each sensor

ts_format
character ["ENVI" | "GDAL" | "Both"] Required output format for virtual file.

out_format
character ["ENVI" | "GTiff"] Format of images used as "input" for the vrt and contained in out_prod_folder/band folders.

verbose
logical If FALSE, suppress processing messages, Default: TRUE

Value

NULL -

Note

License: GPL 3.0