Check if all files required for a given date already exist

Accessory function used to see if all expected out files for the selected date are already present in the output folder. If all expected out files are already present, check_files is set to TRUE, and the date is skipped in MODIStsp_process.

check_files_existence(
  out_prod_folder,
  file_prefix,
  yy,
  DOY,
  bandnames,
  bandsel_orig_choice,
  indexes_bandnames,
  indexes_bandsel,
  quality_bandnames,
  quality_bandsel,
  out_format
)

Arguments

out_prod_folder
character MODIStsp output folder

file_prefix
character File prefix of the processed product (e.g., MOD13Q1)

yy
character year

DOY
character doy

bandnames
character array Bandnames of the MODIS product

bandsel_orig_choice
numeric 0/1 array Indicates which original MODIS layers were selected for processing (does not contain names of bands needed to compute SIs but not selected by the user!)

indexes_bandnames
character array Names of available spectral indexes (standard + custom) available for the currently processed product

indexes_bandsel
numeric 0/1 array Indicates which spectral indexes were selected for processing

quality_bandnames
character array Name of available Quality Indicators for the currently processed product

quality_bandsel
numeric 0/1 array Indicates which Quality Indicators were selected

out_format
character GTiff or ENVI

Value

check - logical = 1 if all expected output files are already existing

Note

License: GPL 3.0