Identify the MODIS original bands needed for a given processing run

Helper function used in MODIStsp_process to identify which MODIS hdf layers are required for the current process. The required layers include all MODIS original layers selected by the user, plus all those required to compute the Spectral Indexes and Quality Indicators selected by the user

get_reqbands(bands_indexes_matrix, indexes_bandsel, indexes_bandnames,
  quality_bandsel, quality_bandnames, out_prod_folder, file_prefix, yy,
  DOY, out_format, reprocess)

Arguments

bands_indexes_matrix
matrix built by set_bandind_matrix

indexes_bandsel
integer 0/1 array array of length equal to the number of Spectral Indexes available for the product (standard + user-provided), set to 1 for indexes to be processed.

indexes_bandnames
character array Abbreviated Names of SIs available for the selected product (used to build output file names of SIs).

quality_bandsel
integer 0|1 array array of length equal to number of available QIs, set to 1 for indexes to be processed.

quality_bandnames
character array Abbreviated Names of Quality Indicators available for the selected product (used to build output file names of QIs).

out_prod_folder
character Main folder where the MODIStsp processed raster will be stored. Used to check if a given processed image already exists.

file_prefix
File prefix corresponding to the MODIS product being processed. Used to check if a given processed image already exists.

yy
Year corresponding to the image being processed. Used to check if a given processed image already exists.

DOY
DOY corresponding to the image being processed. Used to check if a given processed image already exists.. Used to check if a given processed image already exists.

out_format
character ["ENVI" | "GTiff"] Desired output format.

reprocess
character ["Yes" | "No"] If Yes, reprocess data for already existing dates.

Value

req_bands_indexes